Move all tests to test_config.py
